Implemented PSR-6 cache and refactored cache tests
