+++ composer.json
@@ -16,11 +16,13 @@
         "namshi/cuzzle": "^2.0"
     },
     "suggest": {
-        "doctrine/cache": "Allows caching of responses"
+        "doctrine/cache": "Allows caching of responses",
+        "psr/cache": "Allows caching of responses"
     },
     "require-dev": {
         "phpunit/phpunit": "^4.8",
         "doctrine/cache": "^1.1",
+        "psr/cache": "^1.0",
         "symfony/web-profiler-bundle": "^2.3|^3.0",
         "symfony/phpunit-bridge": "^2.7|^3.0"
     },
